# linksdeck-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: LinksDeck Server API
  version: 1.0.0
  description: LinksDeck backend API for mobile app and admin console.
servers:
  - url: /
paths:
...
  /v1/me/developer:
    get:
      operationId: isDeveloper
      summary: Check if current user is a developer/admin
      tags: [me]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Developer check
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsDeveloperResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    IsDeveloperResponse:
      type: object
      required: [isDeveloper]
      properties:
        isDeveloper:
          type: boolean
...